replace cpf with cpf2
